refactor(mcp): extract shared helpers between mcp.ts and mcp-codemode.ts
Both files defined private copies of mcpUserEmail (email resolver),
errorResult (error response shaper), and x-dodo-mcp-depth propagation.
Extract to src/mcp-shared.ts. Per the improve-codebase-architecture
audit — the duplication forced auth and identity-propagation changes to
land in two places.

Behaviour preserved; the two mcpUserEmail copies differed only in their
console.warn label ([mcp] vs [mcp-codemode]). The shared version accepts
an optional label parameter so both call sites retain their original prefix.
